Add unit test program to the cmake build.
Change CPPUNIT_ASSERT_EQUAL functions where comparing literal numU with size_t.
Cast unsigned to size_t.

+++ test/unit/testcal.cpp
@@ -79,11 +79,11 @@ void TestCal::testConstructor()
 {
     Cal::Calendars cal;  // Default constructor with no Schemes.
     Cal::SchemeList slist = cal.get_scheme_list();
-    CPPUNIT_ASSERT_EQUAL( 0U, slist.size() );
+    CPPUNIT_ASSERT_EQUAL( size_t(0), slist.size() );
     CPPUNIT_ASSERT_EQUAL( (Cal::SHandle) NULL, cal.get_scheme( "jdn" ) );
     cal.run_script( "scheme \"jdn\" {name \"Julian Day Number\"; base jdn;}" );
     slist = cal.get_scheme_list();
-    CPPUNIT_ASSERT_EQUAL( 1U, slist.size() );
+    CPPUNIT_ASSERT_EQUAL( size_t(1), slist.size() );
     CPPUNIT_ASSERT_EQUAL( slist[0].handle, cal.get_scheme( "jdn" ) );
 }
